+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+* Added `Polyline.extend` and `Polyline.shorten`.
 ### Added
 
 * Added optional `triangulated` flag to `Mesh.to_vertices_and_faces`.